rename validate_controlled_properties_claims => validate_required_properties_values
