+++ server/models/validations/validate_required_properties_values.js
